+++ lib/src/controllers/drawables/shape/shape_drawables.dart
@@ -3,4 +3,5 @@ export 'line_drawable.dart';
 export 'arrow_drawable.dart';
 export 'rectangle_drawable.dart';
 export 'oval_drawable.dart';
+export 'triangle_drawable.dart';
 export 'double_arrow_drawable.dart';