+++ lib/src/controllers/factories/factories.dart
@@ -3,4 +3,5 @@ export 'line_factory.dart';
 export 'arrow_factory.dart';
 export 'oval_factory.dart';
 export 'rectangle_factory.dart';
+export 'triangle_factory.dart';
 export 'double_arrow_factory.dart';