+++ lib/src/controllers/settings/text_settings.dart
@@ -6,6 +6,9 @@ class TextSettings {
   /// The text style to be used.
   final TextStyle textStyle;
 
+  /// The alignment used to edit and render text.
+  final TextAlign textAlign;
+
   /// Focus node used to edit text.
   /// This focus node will be listened to by the UI to determine user input.
   ///
@@ -13,16 +16,23 @@ class TextSettings {
   /// However, you won't be able to listen to changes in user input focus.
   final FocusNode? focusNode;
 
-  /// Creates a [TextSettings] with the given [textStyle] and [focusNode].
+  /// Creates a [TextSettings] with the given [textStyle], [textAlign] and
+  /// [focusNode].
   const TextSettings({
     this.textStyle = const TextStyle(fontSize: 14, color: Colors.black),
+    this.textAlign = TextAlign.center,
     this.focusNode,
   });
 
   /// Creates a copy of this but with the given fields replaced with the new values.
-  TextSettings copyWith({TextStyle? textStyle, FocusNode? focusNode}) {
+  TextSettings copyWith({
+    TextStyle? textStyle,
+    TextAlign? textAlign,
+    FocusNode? focusNode,
+  }) {
     return TextSettings(
       textStyle: textStyle ?? this.textStyle,
+      textAlign: textAlign ?? this.textAlign,
       focusNode: focusNode ?? this.focusNode,
     );
   }